requirements.txt
pytorch_lightning==1.1.4
transformers==4.30.2
Pillow==8.2.0
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
pyarrow==2.0.0
sacred==0.8.2
pandas==1.1.5
